Refactor parse-method `path'
To address #1589 (Parsers.scala:533) the method `path` is rewritten
to remove the call to `syntaxError`. This does not improve the
error messages per se. However, the cases for `this` and `super`
are now handled a bit more uniformly.

The design choice of using the mutual variable `tTpeName` instead
of passing the `TypeName` as argument to the helper methods is
arguable.